реализовал добавление нового клиента поправил навигацию

+++ app/src/main/java/com/example/seamstress/ui/fragments/clients/listFragment/recycler/ClientsClickListener.kt
@@ -1,5 +1,5 @@
 package com.example.seamstress.ui.fragments.clients.listFragment.recycler
 
 interface ClientsClickListener {
-    fun onClientCardClick(id: Int)
+    fun onClientCardClick(id: Long)
 }
